chore: move modification by formula
