+++ src/services/study/network-modifications.ts
@@ -34,6 +34,7 @@ import {
     LineCreationDto,
     OPERATIONAL_LIMITS_GROUPS_MODIFICATION_TYPE,
     LineModificationDto,
+    ModificationByFormulaDto,
 } from '@gridsuite/commons-ui';
 import { getBaseNetworkModificationUrl, getStudyUrlWithNodeUuid } from './index';
 import { BRANCH_SIDE, OPERATING_STATUS_ACTION } from '../../components/network/constants';
@@ -41,7 +42,6 @@ import type { UUID } from 'node:crypto';
 import {
     AttachLineInfo,
     BalancesAdjustmentInfos,
-    ByFormulaModificationInfos,
     CreateCouplingDeviceInfos,
     CreateVoltageLevelSectionInfos,
     CreateVoltageLevelTopologyInfos,
@@ -1678,7 +1678,7 @@ export function modifyVsc({
 export function modifyByFormula(
     studyUuid: string,
     nodeUuid: UUID,
-    byFormulaModificationInfos: ByFormulaModificationInfos,
+    modificationByFormulaDto: ModificationByFormulaDto,
     uuid?: UUID
 ) {
     let modificationUrl = getNetworkModificationUrl(studyUuid, nodeUuid);
@@ -1696,7 +1696,7 @@ export function modifyByFormula(
             Accept: 'application/json',
             'Content-Type': 'application/json',
         },
-        body: JSON.stringify(byFormulaModificationInfos),
+        body: JSON.stringify(modificationByFormulaDto),
     });
 }
