+++ src/translations/en.json
@@ -8,7 +8,6 @@
 
     "button.delete": "Delete",
     "button.continue": "Continue",
-    "filter": "Filter",
 
     "deleteStudy": "Delete Study",
     "studyNameDidNotMatchMsg": "The given study name did not match with ",
@@ -735,8 +734,6 @@
     "TargetDeadbandMustBeGreaterOrEqualToZero": "Deadband value must be greater than or equal to 0",
     "LowTapPosition": "Low tap position",
     "HighTapPosition": "High tap position",
-    "RatioHighTapPosition": "Ratio high tap position",
-    "PhaseHighTapPosition": "Phase high tap position",
     "TapPosition": "Tap position",
     "HighTapPositionError": "High tap position must be between low tap and 100",
     "TapPositionMustBeBetweenLowAndHighTapPositionValue": "Tap position must be between Low tap position and High tap position",
@@ -829,8 +826,6 @@
     "PrefilledColumn.ReactivePowerLimits": "Reactive power limits",
     "PrefilledColumn.ReactiveCapabilityCurve": "Reactive capability curve",
 
-    "G": "Magnetizing conductance",
-    "B": "Magnetizing susceptance",
     "RatedVoltage": "Rated voltage",
     "TwoWindingsTransformerCreationError": "Error while creating a 2 windings transformer",
     "NewName": "New name",
@@ -1096,17 +1091,8 @@
     "rootNetworkIndexationError": "Root Network indexation process has failed.",
     "checkRootNetworkIndexationError": "Root Network indexation checking has failed.",
     "rootNetworkIndexationStatusUnknown": "Root Network indexation status unknown : \"{status}\".",
-
-    "ReferenceFieldOrValue": "Reference field or value",
-    "Operator": "Operator",
-
     "ModifyByFormula": "Modify by formula",
     "ModifyByAssignment": "Modify by filter",
-    "UseDashToSelectField": "Use # to select a field",
-    "WrongRefOrValueError": "Please enter a valid numeric value or a valid field reference. Use # to select a field",
-    "ValueMustBeNumericWhenPercentageError": "When using %, this field must be a valid positive numeric value",
-    "ValueMustBeRefWhenPercentageError": "When using %, this field must be a valid field reference",
-    "addNewFormula": "Add new formula",
     "waitingRootNetworkIndexation": "Waiting for root network indexation...",
     "undefined": "undefined",
     "GenerationStages": "Stages",
